+++ src/i18n/languages/deDE.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'WENNNV',
     IFS: 'WENNS',
     INDEX: 'INDEX',
+    INDIRECT: 'INDIREKT',
     INT: 'GANZZAHL',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'ZINSZ',