+++ src/i18n/languages/fiFI.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'JOSPUUTTUU',
     IFS: 'JOSS',
     INDEX: 'INDEKSI',
+    INDIRECT: 'EPÄSUORA',
     INT: 'KOKONAISLUKU',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'IPMT',