+++ src/i18n/languages/huHU.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'HAHIÁNYZIK',
     IFS: 'HAELSŐIGAZ',
     INDEX: 'INDEX',
+    INDIRECT: 'INDIREKT',
     INT: 'INT',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'RRÉSZLET',