+++ src/i18n/languages/idID.ts
@@ -123,6 +123,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'JIKA.NA',
     IFS: 'JIKAS',
     INDEX: 'INDEKS',
+    INDIRECT: 'INDIRECT',
     INT: 'BULAT.BAWAH',
     INTERVAL: 'INTERVAL',
     IPMT: 'IPMT',