+++ src/i18n/languages/nbNO.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'HVIS.IT',
     IFS: 'HVIS.SETT',
     INDEX: 'INDEKS',
+    INDIRECT: 'INDIREKTE',
     INT: 'HELTALL',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'RAVDRAG',