+++ src/i18n/languages/nlNL.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'ALS.NB',
     IFS: 'ALS.VOORWAARDEN',
     INDEX: 'INDEX',
+    INDIRECT: 'INDIRECT',
     INT: 'INTEGER',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'IBET',