+++ src/i18n/languages/ruRU.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'ЕСНД',
     IFS: 'УСЛОВИЯ',
     INDEX: 'ИНДЕКС',
+    INDIRECT: 'ДВССЫЛ',
     INT: 'ЦЕЛОЕ',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'ПРПЛТ',