+++ src/i18n/languages/trTR.ts
@@ -122,6 +122,7 @@ const dictionary: RawTranslationPackage = {
     IFNA: 'EĞERYOKSA',
     IFS: 'ÇOKEĞER',
     INDEX: 'İNDİS',
+    INDIRECT: 'DOLAYLI',
     INT: 'TAMSAYI',
     INTERVAL: 'INTERVAL', //FIXME
     IPMT: 'FAİZTUTARI',