+++ src/interpreter/InterpreterState.ts
@@ -11,6 +11,5 @@ export class InterpreterState {
     public formulaAddress: SimpleCellAddress,
     public arraysFlag: boolean,
     public formulaVertex?: FormulaVertex,
-  ) {
-  }
+  ) {}
 }